Fix a bug with precedence
Precedence involving tag counts was broken due to simplifications in the storage of tags and attributes.
Added a test for this and counted the bugs.

On a whim decided that `[@foo<12]` counts slightly more than `.foo` or `[@foo]`, as it is more specific.
Not going to document it yet in case I change my mind.